refactor(parts): split parts.py into parts_core/assets/scan sub-routers (#118)
- Extract 1480-line parts.py into three focused sub-routers:
  parts_core.py (CRUD, stock/lots roll-up, substitutes, meta-members,
  activity), parts_assets.py (asset serving, refresh-from-provider),
  parts_scan.py (bulk-import-from-scan, quick-remove-bag).
- parts_provider.py already existed and is unchanged.
- All four routers registered in main.py under /api/parts — no URL changes.
- Update test_bulk_import_deadline/from_scan/idempotency to import from
  parts_scan instead of the deleted parts module.
- Add test_parts_router_split.py to pin each sub-router registration.
