+++ backend/app/main.py
@@ -109,8 +109,10 @@ from app.api.routes import (
     invitations,
     lots,
     orders,
-    parts,
+    parts_assets,
+    parts_core,
     parts_provider,
+    parts_scan,
     projects,
     reports,
     search,
@@ -362,7 +364,9 @@ _member_gate = [Depends(require_member_for_writes)]
 app.include_router(audit.router, prefix="/api/audit", tags=["audit"], dependencies=_member_gate)
 app.include_router(auth.router, prefix="/api/auth", tags=["auth"])
 app.include_router(workspaces.router, prefix="/api/workspaces", tags=["workspaces"])
-app.include_router(parts.router, prefix="/api/parts", tags=["parts"], dependencies=_member_gate)
+app.include_router(parts_core.router, prefix="/api/parts", tags=["parts"], dependencies=_member_gate)
+app.include_router(parts_assets.router, prefix="/api/parts", tags=["parts"], dependencies=_member_gate)
+app.include_router(parts_scan.router, prefix="/api/parts", tags=["parts"], dependencies=_member_gate)
 app.include_router(storage.router, prefix="/api/storage", tags=["storage"], dependencies=_member_gate)
 app.include_router(stock.router, prefix="/api/stock", tags=["stock"], dependencies=_member_gate)
 app.include_router(lots.router, prefix="/api/lots", tags=["lots"], dependencies=_member_gate)
